added email.js and react router for send emails
