+++ package.json
@@ -3,33 +3,26 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-<<<<<<< HEAD
+    "@abcnews/scrollyteller": "^4.0.2",
     "@testing-library/jest-dom": "^5.11.6",
     "@testing-library/react": "^11.2.2",
     "@testing-library/user-event": "^12.2.2",
+    "emailjs-com": "^2.6.4",
     "intersection-observer": "^0.12.0",
+    "lorem-ipsum": "^2.0.3",
+    "lottie-web": "^5.7.5",
+    "node-sass": "^4.0.0",
     "react": "16.13.1",
     "react-dom": "16.13.1",
     "react-helmet": "6.1.0",
+    "react-lottie": "^1.2.3",
+    "react-router-dom": "^5.2.0",
     "react-scripts": "3.4.3",
     "react-scrollama": "^2.2.15",
     "react-three-fiber": "4.2.21",
     "scrollama": "^2.2.1",
     "three": "0.120.1",
-=======
-    "@abcnews/scrollyteller": "^4.0.2",
-    "@testing-library/jest-dom": "^5.11.4",
-    "@testing-library/react": "^11.1.0",
-    "@testing-library/user-event": "^12.1.10",
-    "lorem-ipsum": "^2.0.3",
-    "lottie-web": "^5.7.5",
-    "node-sass": "^4.0.0",
-    "react": "^17.0.1",
-    "react-dom": "^17.0.1",
-    "react-lottie": "^1.2.3",
-    "react-scripts": "4.0.1",
     "typescript": "^4.1.3",
->>>>>>> bc87589244246c3a9b3bd7bd7190a22037f94708
     "web-vitals": "^0.2.4"
   },
   "scripts": {
